// src/api/types/EnterpriseListResVo.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * 메디스태프 어드민 API & 학회 어드민 API 서버 구성하기
 * 로그인을 제외한 메디스태프 어드민의 모든 API와 학회 관련 API를 구성합니다.
 * OpenAPI spec version: 1.0.0
 */

/**
 * 학회 & 사무국
 */
export interface EnterpriseListResVo {
  adminOpenStatus?: string;
  clientOpenStatus?: string;
  committeeIdx?: number;
  committeeName?: string;
  conferenceEndT?: string;
  conferenceEnterpriseIdx?: number;
  conferenceIdx?: number;
  conferenceName?: string;
  conferencePreRegiEndT?: string;
  conferencePreRegiStartT?: string;
  conferenceRegistrationT?: string;
  conferenceStartT?: string;
  conferenceStringIdx?: string;
  homeUrl?: string;
  isOnline?: string;
  languageType?: string;
  locationIdx?: number;
  locationName?: string;
  needAppAuth?: string;
  templateType?: string;
}